KaligulaTiketsMonitor/requirements.txt
python-telegram-bot==20.8
requests
beautifulsoup4
python-dotenv
telegram